separate service to perform environment cleanup
