+++ api/src/main/java/org/pmiops/workbench/environments/EnvironmentsAdminService.java
@@ -1,0 +1,7 @@
+package org.pmiops.workbench.environments;
+
+import java.util.List;
+
+public interface EnvironmentsAdminService {
+  long deleteUnsharedWorkspaceEnvironmentsBatch(List<String> workspaceNamespaces);
+}
